only re-walk state from current page (rather than start)
