+++ runner/src/server/plugins/engine/models/FormModel.ts
@@ -7,7 +7,6 @@ import {
   FormDefinition,
   Page,
   ConditionRawData,
-  ApiConditionRawData,
   List,
 } from "@xgovformbuilder/model";
 
@@ -17,12 +16,10 @@ import { PageController } from "../pageControllers/PageController";
 import { ExecutableCondition } from "server/plugins/engine/models/types";
 import {
   dateForComparison,
-  getSignature,
   timeForComparison,
 } from "server/plugins/engine/models/FormModel.helpers";
-import wreck from "wreck";
-import hoek from "hoek";
 import pino from "pino";
+import { ApiCondition } from "./ApiCondition";
 
 class EvaluationContext {
   constructor(conditions, value) {
@@ -260,84 +257,5 @@ export class FormModel {
 
   getList(name: string): List | [] {
     return this.lists.find((list) => list.name === name) ?? [];
-  }
-}
-
-class ApiCondition {
-  baseRequest;
-  needsStateEntries;
-  url;
-  logger = pino().child({ name: "apiCondition" });
-
-  constructor(conditionData: ApiConditionRawData) {
-    const { value } = conditionData;
-    const { url, values } = value;
-    this.url = url;
-
-    const entries = Object.entries(values) as [string, string][];
-
-    const { baseRequest, needsStateEntries } = entries.reduce(
-      (prev, [key, value]) => {
-        const needsValueFromState = value.startsWith("$");
-
-        if (needsValueFromState) {
-          // @ts-ignore
-          prev.needsStateEntries.push([key, value.substring(1)]);
-          return prev;
-        }
-
-        prev.baseRequest[key] = value;
-        return prev;
-      },
-      { baseRequest: {}, needsStateEntries: [] }
-    );
-
-    this.baseRequest = baseRequest;
-    this.needsStateEntries = needsStateEntries;
-  }
-
-  requestBody(state) {
-    const dynamicValues = this.needsStateEntries.reduce(
-      (prev, [key, value]) => {
-        prev[key] = hoek.reach(state, value);
-      },
-      {}
-    );
-
-    return {
-      ...this.baseRequest,
-      ...dynamicValues,
-    };
-  }
-
-  makeFn() {
-    return async (state) => {
-      const requestBody = this.requestBody(state);
-      try {
-        return await wreck.post(this.url, {
-          payload: requestBody,
-          headers: {
-            signature: getSignature(JSON.stringify(requestBody)),
-          },
-        });
-      } catch (e) {
-        if (e.isBoom) {
-          this.logger.info({
-            requestBody,
-            response: e.output.payload,
-          });
-          return;
-        }
-        this.logger.error(e);
-      }
-      return false;
-    };
-  }
-
-  static isApiCondition(
-    conditionData: ConditionRawData
-  ): conditionData is ApiConditionRawData {
-    // @ts-ignore
-    return !!conditionData.value?.url;
   }
 }
